pyproject.toml
[tool.poetry]
name = "bicycle-hire-api"
version = "0.1.0"
description = ""
authors = ["dmitriy <dmitriydmgora@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
psycopg = {extras = ["binary"], version = "^3.2.1"}
environs = "^11.0.0"
django-rest-passwordreset = "^1.4.1"
pytest = "^8.2.2"
pytest-django = "^4.8.0"
celery = "^5.4.0"
redis = "^5.0.7"
drf-spectacular = "^0.27.2"
factory-boy = "^3.3.0"
gunicorn = "^22.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
